refactor: 게임 종료 타입에 따른 점수 계산 기능 수정

+++ backend/src/main/java/com/now/naaga/game/domain/Game.java
@@ -168,9 +168,9 @@ public class Game extends BaseEntity {
         throw new GameNotArrivalException(NOT_ARRIVED);
     }
     
-    public BigDecimal findDistance() {
+    public double findDistance() {
         final Position destinationPosition = place.getPosition();
-        return BigDecimal.valueOf(startPosition.calculateDistance(destinationPosition));
+        return startPosition.calculateDistance(destinationPosition);
     }
 
     public Long getId() {
